git_visualizer/requirements.txt
flask==1.1.2
GitPython==3.1.18
google-cloud-bigquery==2.20.0
matplotlib==3.4.2
sqlalchemy==1.4.20